requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
supabase>=2.5.0
pytesseract>=0.3.10
Pillow>=10.0.0
PyMuPDF>=1.24.0
